python interface: support static methods
Currently, a C function is either mapped to a regular method or
to a constructor in the Python interface.
This means that either the first argument needs to be an object
of the class or that the name of the function disappears.
In some cases, it may be useful to preserve the name of the function
even if the first argument is not an object of the class.
This is needed in particular when there are multiple such functions
with the same arguments, such that no distinction could be made
between them in the constructor or when some operation is involved
that is expressed by the name of the function.
For example, the functions isl_val_nan and isl_val_infty cannot
both be converted to (unnamed) constructors because they have
the same arguments.
On the other hand, the function isl_union_map_from_multi_union_pw_aff
should not be converted to an (unnamed) constructor because some
information may be lost in the conversion.

Simply check whether the first argument of an exported function
is an object of the class and declare the method "@staticmethod"
if this is not the case.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>