isl_access_info_compute_flow: extract out access_info_compute_flow_core
isl_access_info_compute_flow internally adds extra dimensions that refer to
the accessed elements and removes them again before returning.
In some cases, it may be useful for the user to know which
array elements(s) are involved in the dependence.
Extract out the part of isl_access_info_compute_flow that does not
project out the extra dimensions so that this part can be reused
to keep the extra dimensions.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>