isl_basic_set_intersect: preserve canonical representation of empty basic sets
The original implementation would just add the constraints of the second
basic set to the first basic set, even if the first basic set is marked
as known to be empty.  The result would be a basic set that is still
marked as known empty, but that no longer has the canonical representation
of such a known empty basic set.

Some parts of the code depend on this canonical representation, or at
least on the fact that there are no such redundant constraints.
In particular, if uset_gist detects that the intersection of
the input and the context is empty, then it returns an empty basic set.
In fact, it returns (the affine hull of) the intersection that was just
found to be empty.  This is fine if the empty set has a canonical
representation, but not if it involves some of the constraints
of the context, since those constraints are not supposed to be
added to the result.  In particular, ever since 9cba0a9 (isl_basic_map_gist:
do not introduce context divs in input, Sun Jun 21 14:19:20 2015 +0200),
isl_basic_map_gist tries to avoid introducing context divs into the input
and any constraints involving those context divs (even if redundant)
will cause problems since they are assumed not to occur in the output
of uset_gist.

Check if the inputs of isl_basic_set_intersect are already obviously
empty.  Technically, this is only needed for the first input, since
that is the one that will be modified, but it does not hurt
to also check whether the other input is obviously empty.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>