isl_ast_build_node_from_schedule: add special case for isolation with no "before"
In case of isolation, the schedule domain is broken up into four parts,
the isolated part, the part that needs to be executed before the isolated
part, the part that needs to be executed after the isolated part
and the remaining elements.
This final part can in principle always be combined with the "before"
or "after" part such that there are only three pieces instead of four,
which in general results in less code being generated.
However, such a combination can be "unnatural" such that
more complicated conditions are generated and their is less opportunity
to simplify expressions inside the generated pieces.
For some applications, this may still be a good trade-off, but
it is not clear if this would be the case for all applications.

However, in the special case where there are no elements that need
to be executed before the isolated part, the risk of combining
the remaining elements with the "wrong" part is much smaller than in general.
Perform such a combination in such cases.
Similar treatment could also be added for the case where there are
no elements that need to be executed after the isolated part.

Requested-by: Tobias Grosser <tobias@grosser.es>
Tested-by: Roman Gareev <gareevroman@gmail.com>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>