isl_ast_codegen.c: add_domain: simplify guard with respect to all constraints
The original code would combine the pending constraints with the guard
derived from the inverse schedule and simplify those with respect
to the generated constraints (without the pending constraints).
However, the pending constraints are guaranteed to be added
afterwards since they cannot get simplified away by any enforced
constraints of the constructed call node.  After all, a call node
does not enforce any constraints.
The pending constraints can therefore be considered as having
been generated and the guard derived from the inverse schedule
can be simplified with respect to the generated constraints
after first having added the pending constraints to the generated constraints.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>