isl_aff.c: plain_pw_multi_aff_from_map: drop constraints involving unknown divs
Unknown divs are not allowed in affine expressions.  Since the local
space of the extracted affine expressions is constructed from the
input basic map, it is safest to remove the unknown divs from this input
first.
This change may not currently have any effect because
plain_pw_multi_aff_from_map is passed an input with only equality constraints,
but this will be changed in an upcoming commit.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>