isl_map_power: fix partitioned Floyd-Warshall implementation
When floyd_warshall_with_groups is used to compute the power relation
(i.e., when "project" is not set), then the space of the elements
in the constructed grid is different from the space of the input.
Initialize the union of the elements in the grid as an empty
map with the correct space.

The incorrect initialization has been there all along,
but the difference in space was ignored until commit
19ff2b8 (isl_map_union_disjoint: check spaces before handling
special cases, Thu Apr 24 11:04:13 2014 +0200).

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>