isl_scheduler.c: compute_split_schedule: splice nested sequences
compute_split_schedule is called when not enough rows can be found
by compute_schedule_wcc.  This function then splits the dependence
graph into two parts in a fairly arbitrary way and then tries again
in each of these two parts.  The results are combined in a sequence.
Inside the two parts, compute_split_schedule may get called again,
resulting in a tree of sequence nodes.
The structure of this tree only reflects how the dependence graph
was broken up internally during the computation of the schedule,
but it does not contain any information that is useful to the user.
Instead, it only serves to confuse the user and it makes it more
difficult to operate on the leaves of this sequence tree.
Flatten this tree into a single sequence.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>