isl_schedule_node_insert_{set,sequence}: simplify elements of set or sequence
In case the input node points to an entire tree, the copies of
these trees that are inserted for each element in the sequence of filters
should be simplified with respect to those elements.
Otherwise, the result may contain a lot of redundant nodes that
can never be reached.

isl_schedule_node_domain_intersect_domain and isl_schedule_node_order_after
already perform a similar simplification.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>