isl_schedule_constraints_compute_schedule: allow merging with unbounded distance
The previous commit introduced the option to incrementally combine SCCs.
However, each time two (or more) clusters of SCCs are combined,
the merge is only allowed if at least one proximity constraint
is completely optimized.

Take for example the atax benchmark from PolyBench:

  for (i = 0; i < _PB_N; i++)
    y[i] = 0;
  for (i = 0; i < _PB_M; i++)
    {
      tmp[i] = SCALAR_VAL(0.0);
      for (j = 0; j < _PB_N; j++)
        tmp[i] = tmp[i] + A[i][j] * x[j];
      for (j = 0; j < _PB_N; j++)
        y[j] = y[j] + A[i][j] * tmp[i];
    }

The third statement is not merged with the fourth statement
because every value written by the third statement is used
by several instance of the fourth statement.  It may however
be useful to tile all statements together.  This would
especially be the case if the third statement were to be removed,
such that there would be a proximity dependence between
the second and the fourth statement.

In both cases, a single value written by one statement is
read by several instances of another statement.  Allow
merging such statements by allowing non-bounded dependence
distances, but only in cases where the proximity constraint
itself does not allow complete optimization in all dimensions.

However, such merges are only allowed after all other complete
merges have been tried.  For example, in atax, the first statement
can also be merged with the fourth statement and this can be done
with zero distances.  This merge should be done first because
after the merge of the third and the fourth statement there
is an internal lower bound on the distance over proximity
constraints and the optimizer will therefore not attempt to
reduce the distance over the proximity constraint between
the first and the fourth statement beyond this lower bound.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>