isl_coalesce.c: add_sub_vars: preserve order of integer divisions
The function add_sub_vars inserts integer divisions that appear in one basic map
into another basic map.  The integer divisions are first appended to the list
of integer divisions and then moved to the right position, if needed.
However, when moving the newly added integer division to its right position,
the integer division at that position was simply moved to the end.
This means that if there is more than one integer division following
the newly inserted integer division, then those integer divisions would
end up in the wrong order.  As a result, the basic map and the tableau
representations would no longer match and no coalescing could be performed.
Preserve the order of the integer divisions while inserting new ones.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>