isl_basic_map_reset_space: also change space if identifiers are different
7ab169d (isl_basic_map_reset_space: special case reset to same space,
Thu Jul 9 13:25:17 2015 +0200) changes isl_basic_map_reset_space to only
modify the input if the space is different from that of the input.
However, isl_basic_map_reset_space may also be used to change
the identifiers of input and/or output dimensions.
In such cases the spaces are considered the same, but the space
of the basic map should still be replaced.

Reported-by: Andreas Kloeckner <lists@informa.tiker.net>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>