isl_map_simplify.c: div_find_coalesce: avoid losing info when coalescing divs
If an existentially quantified variable alpha has bounds l <= alpha < l + m
and always appears together with another existentially quantified variable
beta in the form (alpha + m beta), then this combination is replaced
by a single existentially quantified variable gamma.
Constraints that only contain beta are modified to refer to
beta = (gamma - alpha)/m and alpha is eliminated from the result.

This is a valid transformation if the coefficient of beta is 1 or -1.
For example, assuming l = 0,

	beta <= f

is transformed into

	gamma - alpha <= m f
	gamma <= m f + alpha
	gamma <= m f + m - 1

Given a value for gamma, the corresponding value for beta can then
always be obtained as beta = floor(gamma/m).  Since f is an integral
expression, this value for beta will satisfy the constraint beta <= f.

However, if b has some other coefficient, then the transformation
may result in values of gamma that do not have a corresponding value
for beta.  The resulting map may therefore contain elements that
did not appear in the input.

The div coalescing without the required check on the coefficients was
introduced back in d98e864 (isl_basic_map_drop_redundant_divs: coalesce
divs if possible, Mon May 25 23:43:50 2009 +0200).

Reported-by: Andreas Kloeckner <lists@informa.tiker.net>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>