isl_map_simplify.c: coalesce_divs: fix overly zealous clean-up
Commit 3627737 (isl_map_simplify.c: coalesce_divs: simplify based on fix
in previous commit, Mon Jan 11 13:35:32 2016 +0100) was supposed to
clean up coalesce_divs after the changes made in the previous commit.
However, the code was cleaned up a little too much.
In particular, after the commit elimination would always be performed
using the lower bound, while the lower or upper bound should be used
based on the sign of the corresponding coefficient.

Reported-by: Michael Kruse <meinersbur@googlemail.com>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>