isl_schedule_node_band_split: preserve isolate option
The isolate option was not preserved by the initial implementation,
presumably because for a band there is no way of knowing which
position it has in the schedule tree, while the isolate option
depends on this position.
By passing in the schedule depth to the internal functions,
these function are able to split the isolate option over
the two bands.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>