isl_printer: keep track of notes
This allows the user to keep track of additional state
associated to a printer.
The note is represented by an isl_id * rather than a void *
in order to take advantage of the inherent reference counting.
In particular, the data will be automatically freed in case
isl_id_set_free_user has been called.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>