optionally print AST macro definitions only once
The output of an application that gets printed to a printer
may involve more than one AST node and/or expression.
If so, it is usually sufficient to print the macro definitions
once.  Allow the user to specify that a macro definition should
only be printed once to any given printer, even across multiple
calls to isl_ast_expr_print_macros and/or isl_ast_node_print_macros.

The fact that a macro definition has already been printed
is stored in a note associated to the printer.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>