isl_schedule_node_order_{before,after}: return original node on empty filters
isl_schedule_node_order_before and isl_schedule_node_order_after are
documented as returning the original node when the filter turns out
to be empty.  In case the node is part of a sequence, the parent
would get returned instead.  Fix this to always return the original
node in case of an empty filter.

As a side-effect, the domain that is used to compute the complement
of the filter is now obtained from the original node instead
of the parent (in case the original is part of a sequence).
This reduces the risk of running into an unfiltered extension node
while extracting the domain.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>