isl_basic_map_drop_redundant_divs: fix div with single integral lower bound
If an existentially quantified variable has a single
lower bound with coefficient one that does not involve
any other existentially quantified variables, then the variable
can be safely assigned this value, essentially eliminating
the existentially quantified variable.

Such existentially quantified variables would usually be detected
already by drop_more_redundant_divs, but this test should be cheaper.
Moreover, it gets the case where the coefficient is one out of the way
for when the case of other coefficients will be handled.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>