isl_ast_build_expr_from_pw_aff: optionally detect cases of min/max expressions
An isl_pw_aff with several pieces often represents a minimum of maximum
expression.  Detect some of these cases inside isl_ast_build_expr_from_pw_aff
and construct an explicit minimum or maximum expression on success.
For example construct

	max(0, M)

instead of

	M <= 0 ? 0 : M

Such min/max expressions are usually a lot easier to understand
by humans.  The effect on the compiled code is not entirely clear.
On the one hand, with a min/max expression, there is no need
to evaluate possibly complicated conditions to determine which
of the cases holds.  On the other hand, all expressions in the
sequence now need to be evaluated, while in the expanded case
the first set of conditions may happen to be satisfied, in which
case only the first expression needs to be evaluated.
Also, the domain of the min/max expression becomes the union
of the domains of the individual expressions, meaning that
there is less information that can be used to simplify
the individual expressions.
For example, for an expression

	-32 floor((n + 31)/32)	if n >= 1
	0			if n <= 0

the condition n >= 1 can be exploited to determine
that n + 31 is always non-negative.  However, if the expression
is turned into

	min(-32 floor((n + 31)/32), 0)

then there is no such information that can be exploited.

The detection is performed in a fairly naive way.
The pieces are considered in the order in which they are stored.
If a piece can be combined with the previous piece(s)
to form/extend a min/max expression then such a min/max expression
is constructed/extended.
This will not detect all possible min/max expressions, but it
seems to work fairly well in practice.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>