isl_schedule: store pointer to leaf instead of actual leaf object
The leaf object was embedded in the isl_schedule object presumably
to save out an extra allocation, but this means that this object
cannot be reference counted.  Most of the code is careful
to take into account the special nature of these leaf objects,
but mistakes have happened.
Store a pointer to a leaf object instead so that it can be
treated as any other isl object, reducing the risk for
further such mistakes.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>