extract out shared isl_mat_final_variable_compression
isl_mat_variable_compression and isl_basic_set_variable_compression
contain implementations of variable compression that are nearly
identical to each other, with the second being a slight generalization
of the first.  Extract out a shared isl_mat_final_variable_compression
implementation that is mostly based on the second implementation.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>