isl_basic_map_drop_redundant_divs: exploit equalities among unknown divs
If there are any equality constraints that involve two or more
existentially quantified variables, then the last of these
will be taken as being defined in terms of the others.
This will then cause isl_basic_map_drop_redundant_divs to skip
those existentially quantified variables and they will not be removed.

However, such an equality is actually a perfect way of getting
rid of an existentially quantified variable, either completely
or at least defining an explicit representation for an additional
existentially quantified variable.
In particular, the combination of existentially quantified variables
that appears in the equality constraint can be replaced by
(a multiple of) a single existentially quantified variable
through a unimodular transformation.
For example, for the set

    { [a] : exists e0, e1: 32e1 = 31 + 31a + 31e0 }

the existentially quantified variables e0 and e1 can be replaced
by

	e0' = -31 e0 + 32 e1
	e1' =  -1 e0 +  1 e1

resulting in

    { [a] : exists e0', e1': e0' = 31 + 31a }

e0' can then be plugged into all constraints and removed.
e1' can also be removed because it does not appear in any constraints.
The final result is

    { [a] }

This commit just takes the equalities as they come, but
it might make sense to give priority to equality constraints
where the single existentially quantified variable will appear
with a coefficient one as in that case, it can be eliminated
completely.

Note that normalize_divs called from isl_basic_map_simplify
can have similar effects, but the new code directly targets
this situation.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>