isl_basic_map_preimage_multi_aff: fix handling of integer divisions in strides
The coefficients of the integer divisions were not being copied
to the stride constraints because the number of these integer divisions
was obtained through a call isl_multi_aff_dim(ma, isl_dim_div),
but this always returns zero because the isl_multi_aff itself
does not refer to any integer divisions.  It is only the isl_aff
elements that have integer divisions.  Obtain their number from the caller.

Reported-by: Tobias Grosser <tobias@grosser.es>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>