isl_basic_map_project_out: drop irrelevant constraints first
Constraints that only involve variables that are not related
to the variables that are not projected out cannot influence
the values of those latter variables, except in case they
cause the input to be empty.
Drop these constraints such that the corresponding variables
can be trivially dropped as well.  In general, this results
in fewer remaining existentially quantified variables and/or
less effort spent in trying to discover redundant existentially
quantified variables.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>