move integer division sanity check to isl_basic_map_order_divs
A sanity check that an integer division is not defined in terms
of itself somehow ended up in isl_basic_map_divs_known, presumably
for no other reason than that this function has to visit some
of the integer divisions anyway.
Move the check to isl_basic_map_order_divs instead, where
a failure to satisfy the check would actually result in an infinite loop.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>