isl_mat_variable_compression: reuse local variable
The variables C and TC are used to store the constant term before
and after the transformation, but they are never needed at the same time.
In fact,  The original value is destroyed when the second is computed.
Reuse the same variable to store the two results.
This reduces the difference between the variable compression implementation
in isl_mat_variable_compression and that in
isl_basic_set_variable_compression, which in turn will make it easier
to merge the two.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>