isl_output.c: also print dimension name for dimensions defined by equality
If a dimension is defined by an equality then an affine expression
is printed instead of the dimension name.

For example, the map

	{ [i] -> [j] : j = i }

is printed as

	{ [i] -> [i] }

This has the side effect of dropping the name of the dimension from
the printed representation.

This commit prints both the name and the expression in the form of an equality
if the dimension is named. The behavior is unchanged for unnamed dimensions.

In other words, the same example is printed as follows:

	{ [i] -> [j = i] }

This kind of expression is already parsed correctly by isl.

Signed-off-by: Alexandre Isoard <alexandre.isoard@gmail.com>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>