add AST generation test case test_inputs/codegen/correlation.st
This test case is mainly added to ensure that the generated
code is at least correct since it exercises parts of the AST
generation code that are not exercised by any other AST generation
test case.
However, the schedule in this test case is arguably not very good,
so it is probably not that important to try and produce
nice output for this test case.  In particular, it is important
to keep in mind that there is currently only one such test case
and it is not necessarily representative of the class it belongs to.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>