isl_pw_aff_is_equal: check for NaNs in input
If the inputs involve any NaNs, then they cannot be equal to each other
since a NaN is not considered to be equal to anything.
The explicit check is required because map_from_pw_aff (which
internally calls isl_basic_map_from_aff) cannot be called
on an isl_pw_aff that contains NaNs.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>