add isl_*_to_*_try_get
The functions isl_*_to_*_get essentially assume that a pair with
the given key is available.  In particular, there is no way for
the caller to tell the difference between a missing pair and an error.
This means that the caller always had to call isl_*_to_*_has first.
isl_*_to_*_try_get returns both results, i.e., an isl_bool indicating
whether a pair is present (or an error has occurred) and a copy of
the value if there is such a pair.
This means only a single function needs to be called.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>