isl_map_lexopt_templ.c: basic_map_partial_lexopt*: properly handle NULL empty
While a small part of the code would check for the empty pointer being
NULL, the rest of the code would just assume that this pointer is not NULL.
This is fine because the function was not documented as handling a NULL
pointer and none of the callers pass such a pointer.
However, after the next commit, this function may get called with
a NULL pointer, so add support for that eventuality.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>