isl_map_partial_lexopt_aligned_pw_multi_aff: do not compute empty set when NULL
isl_map_partial_lexopt_aligned_pw_multi_aff is called with a NULL value
for the "empty" pointer from isl_map_lexopt_pw_multi_aff in
isl_map_lexopt_templ.c.  In this case, it makes no sense to compute
the value that would be returned, only to forgot about it at the very end.
Do not compute this set at all if it is not requested by the caller.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>