isl_pw_*_normalize: use isl_sort
isl_pw_*_normalize currently sorts the pieces based on their domains.
Since the domains are disjoint, they are necessarily distinct, so
there is no harm in using qsort.
However, in the next commit, the sorting will be performed
on the basis of the function value expressions and these could
be equal over different pieces.  Different implementations of qsort
may put them in a different order, so use isl_sort instead.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>