isl_basic_map_gist: simplify stride constraints
isl_basic_map_gist already simplifies stride constraints, but only
if they happen to be expressed in exactly the same way in both
the input basic map and the context.
Also simplify other stride constraints by exploiting both
the other equality constraints in the input and all
equality constraints in the context.

For example, for the input

    { [a, b, c] : a mod 6 = 0 and a = c }

and the context

    { [a, b, c] : b mod 2 = 0 and b = c }

before this commit, the stride constraint in the input
would not get simplified because it is expressed using
a different local variable than the one that is used
to express the stride constraint in the context.

With this commit, the other equality constraints in
both input and context are first combined to define
a variable compression

	a = a'
	b = a'
	c = a'

Applying this compression to the stride constraint in
the context yields

	a' mod 2 = 0

which in turn is used to define a parameter compression

	a' = 2 a''

The combined compression is

	a = 2 a''
	b = 2 a''
	c = 2 a''

turning the stride constraint in the input from

	a - 6 alpha = 0

into

	2a'' - 6 alpha = 0

This means that the expression "a", which is supposed to be
a multiple of 6, is equal to "2a''" and is therefore already
known to be a multiple of 2.  The stride constraint can therefore
be simplified to

	a - 3 alpha' = 0

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>