isl_map_partial_lexopt_aligned: compute result as isl_pw_multi_aff first
That is, instead of using a tailored implementation for computing
the result directly as an isl_map, reuse the version that computes
an isl_pw_multi_aff and convert the result to an isl_map.

The tailored implementation would intersect the domains of
the results computed from previous basic maps with the complement of
the domain of the result computed from the next basic map.
This could lead to an explosion in the number of disjuncts
of the final result if the input consists of several disjuncts.
This behavior was illustrated by a large example reported
by Wenlei Bao <guantangbao123@gmail.com>.

Tested-by: Michael Kruse <meinersbur@googlemail.com>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>