isl_schedule_constraints_dump: use YAML facilities of isl_printer
The way an isl_schedule_constraints object is currently printed
is actually in the form of a YAML document already.
However, by using the YAML facilities of the printer, the printing
can easily be changed to flow format instead of block format.
By using an isl_printer to perform the printing, the constituent
isl_union_map objects can now also be printed in standard format
(rather than dump format).  While the dump format of an isl_union_map
can be useful when inspecting individual isl_union_map objects,
it becomes a lot less useful when considering groups of such objects.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>