isl_schedule_node_band_sink: handle expansions
If there are any expansions in the subtree rooted at the input node,
then the domain at the leaves is different from the domain at
the input node.  In order to be able to move this input node
to the leaves, it needs to be expanded according to this expansion.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>