isl_map_unshifted_simple_hull: take into account all translates
map_simple_hull only checks each class of parallel inequality
constraints once (in add_bound).  This works fine in case
of (the shifted) isl_map_simple_hull, because the first representative
of the class is shifted by the required amount in any case (if it turns
out to be an actual bound).  However, in case of isl_map_unshifted_simple_hull,
a more restrictive representative may turn out not to be a bound
on one of the basic sets and then more relaxed representatives would
never be checked.
Replace the representative by its most relaxed translate before
checking whether it is a bound of other basic sets.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>