isl_map_gist: special case subset context
In particular, if the context is a subset of the input, then all
constraints can be removed.  If both the input and the context
consist of a single disjunct, then this situation will already
be discovered, but otherwise, it may have gotten unnoticed.

This helps in particular during AST generation in cases where
the "executed" relation for a particular statement at the leaves
consists of multiple disjuncts.  If the domain constraints are
redundant with respect to the known constraints in the build,
then this could remain unnoticed before, resulting in redundant
conditions in the AST generation output reported by
Tobias Grosser <tobias.grosser@inf.ethz.ch>.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>