isl_tab_detect_redundant: check later constraints first
A given set description may have constraints that are mutually redundant.
That is, either constraint can be removed as redundant, but not both.
This means that the order in which constraints are considered can
lead to different results.  The order in which the constraints are
considered now is difficult to predict by the callers and therefore
also difficult to control.  Check later constraints first such that
later constraints will be end up being removed in favor of earlier
constraints.  This allows the caller to specify which constraints
are preferred by ordering them.

Note that this new selection mechanism results in some fairly random
changes in the AST generation outputs exactly because different constraints
may end up getting preserved.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>