isl_map_*_simple_hull: sort constraints before dropping redundant constraints
Various heuristics in isl_pw_multi_aff_from_map are based on investigating
the constraints of the unshifted simple hull.  In particular,
one of the heuristics checks for pairs of constraints that define
an integer division.  There may, however, be different possible
such pairs of constraints in the original input that could
both be valid for the entire input, but that may be redundant
with respect to each other.  In such cases, it is possible
for one constraint of both pairs to end up getting removed
such that no pair is preserved.  For example, the set
of all valid constraints collected by map_simple_hull may be

{ [a, b] : 0 <= a <= 127 and -118 - a <= 127b <= -a and -119 - a <= 128b <= -a }

where there are two pairs of integer division defining constraints on b,
but the removal of redundant constraints may simplify this to

{ [a, b] : 0 <= a <= 127 and -118 - a <= 127b and 128b <= -a }

where no such pair is left.
Sort the constraints before looking for redundant constraints
such that the constraints in one pair are considered before
the constraints in the other pair, increasing the chance
that at least one pair will remain.

Perhaps the heuristics in isl_pw_multi_aff_from_map should consider
all valid constraints in the disjuncts of the input instead of
only some non-redundant set of such constraints, but this is left
for later.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>